WIP: remove `src/` folder, move all used files to `lib/`
