+++ lib/add-contributor.js
@@ -1,7 +1,7 @@
 module.exports = addContributor;
 
-const getUserDetails = require("../src/tasks/processIssueComment/utils/getUserDetails");
-const ContentFiles = require("../src/tasks/processIssueComment/ContentFiles");
+const getUserDetails = require("./get-user-details");
+const ContentFiles = require("./content-files");
 
 async function addContributor({
   context,
