+++ lib/get-user-details.js
@@ -1,3 +1,5 @@
+module.exports = getUserDetails;
+
 const { UserNotFoundError } = require("./errors");
 
 async function getUserDetails({ octokit, username }) {
@@ -30,5 +32,3 @@ async function getUserDetails({ octokit, username }) {
     profile: blog || html_url,
   };
 }
-
-module.exports = getUserDetails;
