+++ lib/is-message-for-bot.js
