+++ lib/repository.js
@@ -2,7 +2,7 @@ const {
   AllContributorBotError,
   BranchNotFoundError,
   ResourceNotFoundError,
-} = require("../utils/errors");
+} = require("./errors");
 
 class Repository {
   constructor({ repo, owner, octokit, defaultBranch, log }) {
