+++ lib/setup-options-config.js
@@ -1,9 +1,7 @@
 module.exports = setupOptionsConfig;
 
-const OptionsConfig = require("../src/tasks/processIssueComment/OptionsConfig");
-const {
-  ResourceNotFoundError,
-} = require("../src/tasks/processIssueComment/utils/errors");
+const OptionsConfig = require("./options-config");
+const { ResourceNotFoundError } = require("./errors");
 
 async function setupOptionsConfig({ repository }) {
   const optionsConfig = new OptionsConfig({
