+++ lib/setup-repository.js
@@ -1,9 +1,7 @@
 module.exports = setupRepository;
 
-const Repository = require("../src/tasks/processIssueComment/Repository");
-const {
-  BranchNotFoundError,
-} = require("../src/tasks/processIssueComment/utils/errors");
+const Repository = require("./repository");
+const { BranchNotFoundError } = require("./errors");
 
 async function setupRepository({ context, branchName }) {
   const defaultBranch = context.payload.repository.default_branch;
